Improved the possible word usage selection.
Removes words that include letters matching the pattern but having
letters that are not in the base set or included pattern letters.

Does not take into account the number of letters however. So abc as the
valid letters will say that abba is a valid word because b exists.
Further improvement would exclude that since only 1 b exists and 2 are
needed.